# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FREESPORT API
  version: 1.0.0
  x-logo:
    url: https://api.freesport.ru/static/logo.png
    altText: FREESPORT API
  description: RESTful API для B2B/B2C платформы спортивных товаров FREESPORT
paths:
...
  /products/visible-brands/:
    get:
      operationId: products_visible_brands_retrieve
      description: Возвращает ID брендов, у которых есть товары по текущим фильтрам
        каталога. Параметр brand игнорируется — endpoint отражает глобальные фильтры,
        не сужая sidebar до уже выбранных брендов.
      summary: Видимые бренды по фильтрам
      parameters:
      - in: query
        name: category_id
        schema:
          type: integer
        description: ID категории (включая дочерние)
      - in: query
        name: in_stock
        schema:
          type: boolean
        description: Товары в наличии
      - in: query
        name: max_price
        schema:
          type: number
        description: Максимальная цена
      - in: query
        name: min_price
        schema:
          type: number
        description: Минимальная цена
      - in: query
        name: search
        schema:
          type: string
        description: Поисковый запрос
      tags:
      - Products
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
          description: ''
  /products/visible-categories/:
    get:
      operationId: products_visible_categories_retrieve
      description: Возвращает ID категорий, в которых есть товары, соответствующие
        активным фильтрам. Параметр category_id игнорируется — endpoint отражает глобальные
        фильтры, не сужая sidebar до активной ветки.
      summary: Видимые категории по фильтрам
      parameters:
      - in: query
        name: brand
        schema:
          type: string
        description: Бренд (ID или slug)
      - in: query
        name: in_stock
        schema:
          type: boolean
        description: Товары в наличии
      - in: query
        name: max_price
        schema:
          type: number
        description: Максимальная цена
      - in: query
        name: min_price
        schema:
          type: number
        description: Минимальная цена
      - in: query
        name: search
        schema:
          type: string
        description: Поисковый запрос
      tags:
      - Products
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
          description: ''
...
components:
  schemas:
...
    Brand:
      type: object
      description: Serializer для брендов
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: Название бренда
          maxLength: 100
        slug:
          type: string
          pattern: ^[-a-zA-Z0-9_]+$
        image:
          type:
          - string
          - 'null'
          description: Возвращает URL изображения или None.
          readOnly: true
        description:
          type: string
          title: Описание
        website:
          type: string
          format: uri
          title: Веб-сайт
          maxLength: 200
        is_featured:
          type: boolean
          title: Показывать на главной
      required:
      - id
      - image
      - name
...
    ProductList:
      type: object
      description: |-
        Serializer для списка товаров (базовая модель Product)

        Примечание: Product - это базовая модель товара без цен и остатков.
        Цены и остатки хранятся в ProductVariant.
        Для получения вариантов используйте вложенное поле 'variants'
        в ProductDetailSerializer.

        Вычисляемые поля для обратной совместимости (данные из вариантов):
        - retail_price, opt1_price, opt2_price, opt3_price: цены из первого варианта
        - stock_quantity: суммарное количество на складе по всем вариантам
        - is_in_stock: есть ли хотя бы один вариант в наличии
        - main_image: изображение из первого варианта или base_images
        - can_be_ordered: можно ли заказать товар
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: Название
          maxLength: 300
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        brand:
          allOf:
          - $ref: '#/components/schemas/Brand'
          readOnly: true
        category:
          type: string
          readOnly: true
        description:
          type: string
          title: Описание
        short_description:
          type: string
          title: Краткое описание
          maxLength: 500
        specifications:
          readOnly: true
        base_images:
          title: Базовые изображения
          description: Общие изображения товара из 1С (используются как fallback для
            вариантов)
        is_featured:
          type: boolean
          title: Рекомендуемый
        is_hit:
          type: boolean
          readOnly: true
          title: Хит продаж
          description: Отображать бейдж 'Хит продаж' на карточке товара
        is_new:
          type: boolean
          readOnly: true
          title: Новинка
          description: Отображать бейдж 'Новинка' на карточке товара
        is_sale:
          type: boolean
          readOnly: true
          title: Распродажа
          description: Товар участвует в распродаже
        is_promo:
          type: boolean
          readOnly: true
          title: Акция
          description: Товар участвует в акции/промо
        is_premium:
          type: boolean
          readOnly: true
          title: Премиум
          description: Премиум товар (эксклюзив, лимитированная серия)
        discount_percent:
          type:
          - integer
          - 'null'
          readOnly: true
          title: Процент скидки
          description: Процент скидки для отображения на бейдже (0-100)
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Дата создания
        attributes:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: |-
            Получить атрибуты продукта

            Args:
                obj: Product instance

            Returns:
                list[dict]: Список атрибутов с полями name, value, slug, type
          readOnly: true
        retail_price:
          type: number
          format: double
          description: Получить розничную цену из первого варианта
          readOnly: true
        opt1_price:
          type: number
          format: double
          description: Получить оптовую цену уровня 1 из первого варианта
          readOnly: true
        opt2_price:
          type: number
          format: double
          description: Получить оптовую цену уровня 2 из первого варианта
          readOnly: true
        opt3_price:
          type: number
          format: double
          description: Получить оптовую цену уровня 3 из первого варианта
          readOnly: true
        stock_quantity:
          type: integer
          description: Получить суммарное количество на складе по всем вариантам
          readOnly: true
        is_in_stock:
          type: boolean
          description: Проверить наличие товара (любой вариант в наличии)
          readOnly: true
        main_image:
          type:
          - string
          - 'null'
          description: |-
            Получить основное изображение из первого варианта или base_images

            Возвращает относительный URL с /media/ префиксом
          readOnly: true
        can_be_ordered:
          type: boolean
          description: Проверить можно ли заказать товар
          readOnly: true
        current_price:
          type: string
          description: Получить актуальную цену на основе роли пользователя
          readOnly: true
        sku:
          type: string
          description: Получить артикул первого варианта
          readOnly: true
        rrp:
          type:
          - number
          - 'null'
          format: double
          description: Получить РРЦ первого варианта
          readOnly: true
        msrp:
          type:
          - number
          - 'null'
          format: double
          description: Получить МРЦ первого варианта
          readOnly: true
      required:
      - attributes
      - brand
      - can_be_ordered
      - category
      - created_at
      - current_price
      - description
      - discount_percent
      - id
      - is_hit
      - is_in_stock
      - is_new
      - is_premium
      - is_promo
      - is_sale
      - main_image
      - msrp
      - name
      - opt1_price
      - opt2_price
      - opt3_price
      - retail_price
      - rrp
      - sku
      - slug
      - specifications
      - stock_quantity
...
servers:
- url: http://127.0.0.1:8001
  description: Development server
- url: https://api.freesport.ru
  description: Production server
...